ImageFrequency/requirements.txt
###### Requirements without Version Specifiers ######`
numpy
pandas
scipy
Pillow
